test(ui): keep browser regressions within lint boundaries
